Enforce that cached include objects are properly expanded.
This patch fixes an inefficiency of the implementation of Include
objects. Namely, each time an Include has to be imported, its content
is recomputed on the fly by performing substitution of the included
module. We could force that all such objects are always expanded,
but this would defeat the purpose of the indirection, which is
to keep a compact representation of included modules in object files.

Rather, we generalize the type of objects so as to have two distinct
representations, one for the object file and one for the Rocq process.
This allows enforcing that the module cache used by Declaremods only
keeps substituted objects by typing, while maintaining the indirect
form for object files. This patch may augment memory consumption to
some extent since we now cache a larger object, but it is likely
that most of this memory was already retained because all objects
contained in the included modules were added to the summary.