Small memory optimization in Nametab.
Instead of duplicating the same data when pushing it several times in
a row, we check that the head value is not equal to the one being
added. This does not solve the issue that the Nametab structure
is fundamentally leaky when overwriting bindings, but in practice
it seems to mitigate it well enough.