Merge the Dumpglob API to start dumping and to push a glob output.
All users were synchronized, except that the main entry point was forgetting
to end the dumping, since it would exit prematurely. Furthermore, we clean
up along the way the internal states of the Dumpglob module. I am not sure
this code was really designed this way, it looks like its dire state is
the result of organic modification, but this resulted in very dubious
bits of code.