Add a cache to Dumpglob output.
Due to the way Ltac runtime is implemented, if a constant is
evaluated during a tactic call, it will generate a glob output.
In particular, if it is evaluated several times, the glob file will
contain the same line repeated again and again, leading to
potentially huge glob files.

For all consumers, this repetition is useless because they will
just overwrite the same data again and again. Said otherwise, clients
are idempotent w.r.t. repeated declarations. We leverage this
property by simply caching whether a specific reference has already
been registered in the glob file and ignoring it otherwise.